# api/openapi/accounts.yaml
openapi: 3.0.4
info:
  title: Banking - Accounts
  description: This is an accounts service API based on the OpenAPI 3.0 specification.
  version: 1.0.0
components:
  schemas:
    CustomerLogin:
      type: string
      example: ipet123
    CustomerForename:
      type: string
      example: Ivan
    CustomerSurname:
      type: string
      example: Petrov
    Customer:
      type: object
      properties:
        login:
          $ref: '#/components/schemas/CustomerLogin'
        forename:
          $ref: '#/components/schemas/CustomerForename'
        surname:
          $ref: '#/components/schemas/CustomerSurname'
        birthdate:
          type: string
          format: date
          example: 1984-04-17
      required:
        - login
        - forename
        - surname
        - birthdate
    Currency:
      $ref: './common.yaml#/components/schemas/Currency'
    CurrencyAmount:
      $ref: './common.yaml#/components/schemas/CurrencyAmount'
    AccountNumber:
      type: integer
      format: int64
      example: 438832491231
    Account:
      type: object
      properties:
        number:
          $ref: '#/components/schemas/AccountNumber'
        currency:
          $ref: '#/components/schemas/Currency'
        amount:
          $ref: '#/components/schemas/CurrencyAmount'
      required:
        - number
        - currency
        - amount
  parameters:
    CustomerLogin:
      name: login
      description: Customer login.
      required: true
      in: path
      schema:
        $ref: '#/components/schemas/CustomerLogin'
    AccountNumber:
      name: accountNumber
      description: An account number.
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/AccountNumber'
    IdempotencyKey:
      name: Idempotency-Key
      description: UUID of mutable operation.
      in: header
      required: true
      schema:
        type: string
        format: uuid
        example: f81d4fae-7dec-11d0-a765-00a0c91e6bf6
  responses:
    ErrorMessage:
      $ref: './common.yaml#/components/responses/ErrorMessage'
    AccountOperation:
      description: Amount of money in the account after the operation.
      content:
        application/json:
          schema:
            title: AccountOperationResponse
            type: object
            properties:
              currency:
                $ref: '#/components/schemas/Currency'
              totalAmount:
                $ref: '#/components/schemas/CurrencyAmount'
            required:
              - currency
              - totalAmount
paths:
  /customers:
    get:
      tags:
        - customers
      operationId: getCustomers
      summary: Get all customers.
      description: Get information about all customers.
      responses:
        '200':
          description: List of customers.
          content:
            application/json:
              schema:
                type: array
                items:
                  title: CustomerNames
                  type: object
                  properties:
                    login:
                      $ref: '#/components/schemas/CustomerLogin'
                    forename:
                      $ref: '#/components/schemas/CustomerForename'
                    surname:
                      $ref: '#/components/schemas/CustomerSurname'
                  required:
                    - login
                    - forename
                    - surname
        '400':
          $ref: '#/components/responses/ErrorMessage'
    post:
      tags:
        - customers
      operationId: createCustomer
      summary: Create customer.
      description: Create a new customer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer created.
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /customers/{login}/info:
    get:
      tags:
        - customers
      operationId: getCustomerInfo
      summary: Get customer information.
      description: Get information about customer - personal data and accounts.
      parameters:
        - $ref: '#/components/parameters/CustomerLogin'
      responses:
        '200':
          description: Customer information.
          content:
            application/json:
              schema:
                title: CustomerInfo
                type: object
                properties:
                  customer:
                    $ref: '#/components/schemas/Customer'
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
                required:
                  - customer
                  - accounts
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /customers/{login}/balance:
    get:
      tags:
        - customers
      operationId: getCustomerBalance
      summary: Get customer balance.
      description: Get the customer balance over all accounts in the specified currency.
      parameters:
        - $ref: '#/components/parameters/CustomerLogin'
        - name: currency
          description: Currency code for the balance representation.
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/Currency'
      responses:
        '200':
          description: Customer balance.
          content:
            application/json:
              schema:
                title: BalanceResponse
                type: object
                properties:
                  currency:
                    $ref: '#/components/schemas/Currency'
                  amount:
                    $ref: '#/components/schemas/CurrencyAmount'
                required:
                  - currency
                  - amount
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /accounts:
    post:
      tags:
        - accounts
      operationId: createAccount
      summary: Create account.
      description: Create a new customer account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: CreateAccountRequest
              type: object
              properties:
                customerLogin:
                  $ref: '#/components/schemas/CustomerLogin'
                currency:
                  $ref: '#/components/schemas/Currency'
              required:
                - customerLogin
                - currency
      responses:
        '200':
          description: Created account number.
          content:
            application/json:
              schema:
                title: CreateAccountResponse
                type: object
                properties:
                  number:
                    $ref: '#/components/schemas/AccountNumber'
                required:
                  - number
        '400':
          $ref: '#/components/responses/ErrorMessage'
    get:
      tags:
        - accounts
      operationId: getAccountNumberByCustomer
      summary: Get account number by customer data.
      description: Get account number by customer login and currency.
      parameters:
        - name: login
          description: Customer login.
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/CustomerLogin'
        - name: currency
          description: Account currency.
          required: true
          in: query
          schema:
            $ref: '#/components/schemas/Currency'
      responses:
        '200':
          description: Account number.
          content:
            application/json:
              schema:
                title: AccountNumberResponse
                type: object
                properties:
                  number:
                    $ref: '#/components/schemas/AccountNumber'
                required:
                  - number
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /accounts/{accountNumber}:
    get:
      tags:
        - accounts
      operationId: getAccountInfo
      summary: Get account information.
      description: Get information about the account - customer login, currency, amount.
      parameters:
        - $ref: '#/components/parameters/AccountNumber'
      responses:
        '200':
          description: Information about the account.
          content:
            application/json:
              schema:
                title: AccountInfo
                type: object
                properties:
                  customerLogin:
                    $ref: '#/components/schemas/CustomerLogin'
                  currency:
                    $ref: '#/components/schemas/Currency'
                  amount:
                    $ref: '#/components/schemas/CurrencyAmount'
                required:
                  - customerLogin
                  - currency
                  - amount
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /accounts/{accountNumber}/top-up:
    put:
      tags:
        - accounts
      operationId: topUpAccount
      summary: Top up account.
      description: Top up the account by the passed amount.
      parameters:
        - $ref: '#/components/parameters/AccountNumber'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: TopUpAccountRequest
              type: object
              properties:
                augend:
                  $ref: '#/components/schemas/CurrencyAmount'
              required:
                - augend
      responses:
        '200':
          $ref: '#/components/responses/AccountOperation'
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /accounts/{accountNumber}/withdraw:
    put:
      tags:
        - accounts
      operationId: withdrawAccount
      summary: Withdraw account.
      description: Withdraw money from the account.
      parameters:
        - $ref: '#/components/parameters/AccountNumber'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: WithdrawAccountRequest
              type: object
              properties:
                subtrahend:
                  $ref: '#/components/schemas/CurrencyAmount'
              required:
                - subtrahend
      responses:
        '200':
          $ref: '#/components/responses/AccountOperation'
        '400':
          $ref: '#/components/responses/ErrorMessage'
  /transfer:
    put:
      tags:
        - transfer
      operationId: transfer
      summary: Transfer money.
      description: Transfer money from the account to the another account.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              title: TransferRequest
              type: object
              properties:
                senderAccountNumber:
                  $ref: '#/components/schemas/AccountNumber'
                receiverAccountNumber:
                  $ref: '#/components/schemas/AccountNumber'
                senderCurrencyAmount:
                  $ref: '#/components/schemas/CurrencyAmount'
              required:
                - senderAccountNumber
                - receiverAccountNumber
                - senderCurrencyAmount
      responses:
        '200':
          description: Successful transfer.
          content:
            application/json:
              schema:
                title: TransferResponse
                type: object
                properties:
                  totalSenderAmount:
                    $ref: '#/components/schemas/CurrencyAmount'
                  totalReceiverAmount:
                    $ref: '#/components/schemas/CurrencyAmount'
                required:
                  - totalSenderAmount
                  - totalReceiverAmount
        '400':
          $ref: '#/components/responses/ErrorMessage'